[fix] Make sure the slides are sorted
`-O` is used as a quick and dirty way to make sure the images are easily
sortable. Previously, the counting number was burried in the image's file name.

Since wget's `-P` argument does not work in combination with `-O`,
the directory for the downloaded images has to be created by us.